refactor: extract layout to a hook
