+++ gbajs3/src/components/modals/upload-rom.tsx
@@ -9,7 +9,6 @@ import { ModalBody } from './modal-body.tsx';
 import { ModalFooter } from './modal-footer.tsx';
 import { ModalHeader } from './modal-header.tsx';
 import { EmulatorContext } from '../../context/emulator/emulator.tsx';
-import { LayoutContext } from '../../context/layout/layout.tsx';
 import { ModalContext } from '../../context/modal/modal.tsx';
 import {
   EmbeddedProductTour,
@@ -48,7 +47,6 @@ const BiCloudUploadLarge = styled(BiCloudUpload)`
 export const UploadRomModal = () => {
   const { setIsModalOpen } = useContext(ModalContext);
   const { emulator } = useContext(EmulatorContext);
-  const { hasSetLayout, clearLayouts } = useContext(LayoutContext);
   const {
     register,
     reset,
@@ -80,7 +78,6 @@ export const UploadRomModal = () => {
         emulator.filePaths().gamePath + '/' + romFile.name
       );
       if (hasSucceeded) {
-        if (!hasSetLayout) clearLayouts();
         setIsModalOpen(false);
       }
     };
